remove AspNet from method/class names for the .net adapters. Add check to call AddTechTolkLocalizationAdapters() AFTER Microsoft's AddLocalization()
