+++ src/TechTolk.Extensions.Localization/TechTolkServiceCollectionExtensions.cs
@@ -7,9 +7,22 @@ namespace TechTolk;
 
 public static class TechTolkServiceCollectionExtensions
 {
-    public static void AddTechTolkAspNetLocalization(this IServiceCollection services)
+    public static IServiceCollection AddTechTolkLocalizationAdapters(this IServiceCollection services)
     {
-        services.AddLocalization();
-        services.Replace(ServiceDescriptor.Singleton<IStringLocalizerFactory, TechTolkStringLocalizerFactory>());
+        var factoryType = typeof(IStringLocalizerFactory);
+
+        if (services.Any(s => s.ServiceType == factoryType))
+        {
+            services.Replace(ServiceDescriptor.Singleton<IStringLocalizerFactory, TechTolkStringLocalizerFactory>());
+        }
+        else
+        {
+            throw new InvalidOperationException(
+                $"Unable to replace the {nameof(ServiceDescriptor)} of service type {factoryType.Name}, " +
+                $"because no such descriptor is registered. \n" +
+                $"Did you call {nameof(AddTechTolkLocalizationAdapters)} *after* calling {nameof(LocalizationServiceCollectionExtensions.AddLocalization)}?");
+        }
+
+        return services;
     }
 }
